# specs/api/openapi_v0.1.yaml
openapi: "3.1.0"
info:
  title: MindLakeOS Demo API
  version: "0.1"
  description: |
    MindLakeOS Demo（Doc2Skill 认知编译平台）的 REST API 合同。
    
    覆盖四个核心端点：
    - `POST /compile` — 文档编译，生成 SemanticKernelIR + Skill 目录包
    - `GET /skills` — 列出所有已编译 Skill
    - `GET /skills/{id}` — 获取单个 Skill 详情
    - `POST /run` — 挂载 Skill 执行用户输入，返回 trace/evidence/violations
    - `POST /bench` — 运行对比评测（Vanilla vs RAG vs MindLakeOS）
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
...
paths:
...
  /skills/{id}:
    get:
      tags: [skills]
      operationId: getSkill
      summary: 获取单个 Skill 详情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Skill ID（kebab-case），如 `ops-sev1-incident-response`
      responses:
        "200":
          description: Skill 详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillDetail"
        "404":
          description: Skill 不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    SkillSummary:
      type: object
      required: [skill_id, skill_name, description, domain, object_type, compiled_at]
      properties:
        skill_id:
          type: string
        skill_name:
          type: string
        description:
          type: string
        domain:
          type: string
        object_type:
          type: string
        compiled_at:
          type: string
          format: date-time
    SkillDetail:
      allOf:
        - $ref: "#/components/schemas/SkillSummary"
        - type: object
          required: [ir, files_tree]
          properties:
            ir:
              $ref: "#/components/schemas/SemanticKernelIR"
            files_tree:
              type: array
              items:
                $ref: "#/components/schemas/SkillFile"
    SemanticKernelIR:
      type: object
      required: [kernel_id, version, r, n, e, t]
      description: R/N/E/T 四核中间表示，详见 specs/rnet/ir_schema_v0.3.json
      properties:
        kernel_id:
          type: string
        version:
          type: string
        compiled_at:
          type: string
          format: date-time
        source_doc:
          type: string
        compilation_strategy:
          type: string
          enum: [minimal, default, detailed]
        r:
          type: object
          description: R 核 — 详见 ir_schema_v0.3.json#/$defs/RCore
        n:
          type: object
          description: N 核 — 详见 ir_schema_v0.3.json#/$defs/NCore
        e:
          type: object
          description: E 核 — 详见 ir_schema_v0.3.json#/$defs/ECore
        t:
          type: object
          description: T 核 — 详见 ir_schema_v0.3.json#/$defs/TCore
...
    SkillFile:
      type: object
      required: [path, type]
      properties:
        path:
          type: string
        type:
          type: string
          enum: [skill_main, reference, asset, script, test]
        size_bytes:
          type: integer
        token_estimate:
          type: integer
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: 错误代码，如 INVALID_REQUEST / COMPILE_FAILED / SKILL_NOT_FOUND
        message:
          type: string
          description: 人类可读的错误说明
        details:
          type: object
          description: 附加错误信息（可选）